Move common parts of *MapWrappers to abstract base class
